% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_of_subtype.R
\name{fequency_of_subtype}
\alias{fequency_of_subtype}
\title{Compute the frequency of subtype prediction}
\usage{
fequency_of_subtype(SPI)
}
\arguments{
\item{SPI}{A tibble with columns sample, subtype_prediction, index, as produced
by \code{make_subtype_predictions}}
}
\value{
A tibble with columns sample, subtype_prediction, frequency. We call
this object SPF, for \emph{subtype prediction frequency}
}
\description{
This function computes, for each sample and each possible subtype, the fraction of
kTSP models in the family that predict this subtype assignment
}
