% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_prediction.R
\name{identify_mpred_by_model}
\alias{identify_mpred_by_model}
\title{Identifies the Multi-kTSP model object corresponding to the \code{model}
parameter}
\usage{
identify_mpred_by_model(model)
}
\arguments{
\item{model}{A character string that is one of "erposneg", "erpos", "erneg",
to determine whether BCCSclassifier(ER+/-), BCCSclassifier(ER+), or
BCCSclassifier(ER-), respectively, should be used for subtyping.}
}
\value{
The multi-kTSP model used for prediction (termed an MPred object for short)
corresponding to the input \code{model}
}
\description{
Identifies the Multi-kTSP model object corresponding to the \code{model}
parameter
}
