% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_prediction.R
\name{make_pairwise_predictions}
\alias{make_pairwise_predictions}
\title{Predict subtypes, pairwise and by index}
\usage{
make_pairwise_predictions(dat, model)
}
\arguments{
\item{dat}{An object of class matrix with samples as column names and gene symbols
as row names, and entries the corresponding expression levels. The included tibble
\code{bccs_classifier_genes} contains the gene symbols and ENTREZIDs used in the
BCCSclassifier models, along with Affymetrix (hgu133a) and Illumina
(illuminaHuman v3) probes that can be used to represent the associated genes.
Expression values must \bold{not} be scaled or median-centered.}

\item{model}{A character string that is one of "erposneg", "erpos", "erneg",
to determine whether BCCSclassifier(ER+/-), BCCSclassifier(ER+), or
BCCSclassifier(ER-), respectively, should be used for subtyping.}
}
\value{
The output is a tibble (data.frame),with columns sample, pair,
pair_prediction, index. pair indicates the pair of subtypes tested in the
kTSP inner model. pair_prediction is the subtype in the pair chosen
by the kTSP model. index (an integer) is the index of the kTSP model in the
Multi-kTSP family. This object is called a PPI object, short for
\emph{pairwise prediction - indexed}
}
\description{
This is the most basic step in the process of predicting a subtype for a
sample. The BCCSclassifier multi-kTSP object is a list of kTSP models, each
consisting of a list of kTSP tests for the best subtype between a pair of possible
choices, ranging over all kTSP models in the family. This
function applies ktspair prediction to each of these inner models and arranges
the results in a tibble, respecting the kTSP model index.
}
\details{
A function that tests, for each sample, each kTSP model in the family,
and each possible pair of subtypes, which subtype is the best choice for the sample
}
