% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype_prediction.R
\name{make_subtype_predictions}
\alias{make_subtype_predictions}
\title{Predict subtype assignments, by index}
\usage{
make_subtype_predictions(PPI)
}
\arguments{
\item{PPI}{A tibble with columns sample, pair, pairwise_prediction, index,
as produced by the \code{make_pairwise_predictions} function.}
}
\value{
A tibble with columns sample, subtype_prediction, index, that we
consider an SPI object, short for \emph{subtype prediction - indexed}
}
\description{
This function selects the subtype that wins the most pairwise competitions, for each
kTSP model in the family
}
\details{
Predicts, for each sample and kTSP model, the best subtype using the
pairwise test results in the given kTSP model
}
