% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_bccs_frequency}
\alias{predict_bccs_frequency}
\title{Frequency of BCCS subtype prediction}
\usage{
predict_bccs_frequency(dat, model)
}
\arguments{
\item{dat}{An object of class matrix with samples as column names and gene symbols
as row names, and entries the corresponding expression levels. The included tibble
\code{bccs_classifier_genes} contains the gene symbols and ENTREZIDs used in the
BCCSclassifier models, along with Affymetrix (hgu133a) and Illumina
(illuminaHuman v3) probes that can be used to represent the associated genes.
Expression values must \bold{not} be scaled or median-centered.}

\item{model}{A character string that is one of "erposneg", "erpos", "erneg",
to determine whether BCCSclassifier(ER+/-), BCCSclassifier(ER+), or
BCCSclassifier(ER-) should be used for subtyping.}
}
\value{
The output is a tibble (data.frame) with columns sample, subtype_prediction,
frequency. There is one record for each sample in the column names of \code{dat} and
each possible subtype. \code{frequency} is the fraction of kTSP models in
the family that predicted the sample was in the corresponding subtype.
}
\description{
A function to predict for each sample and subtype, the fraction of
kTSP models in the family that predict the sample is in the subtype.

Learn more in \code{vignette("BCCSclassifier")}
}
