% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_subtype_prediction.R
\name{select_final_prediction}
\alias{select_final_prediction}
\title{BCCS prediction from frequency}
\usage{
select_final_prediction(SPF)
}
\arguments{
\item{SPF}{A tibble with columns sample, subtype_prediction, frequency, as
produced by \code{fequency_of_subtype}.}
}
\value{
A tibble with columns sample, subtype_prediction, frequency, with one
record per sample
}
\description{
This function identifies, for each sample, the subtype that is predicted
most frequently across the multi-kTSP family, or reports a "tie" if the
maximum isn't unique. The result returns the maximum frequency as a column
in the output.
}
